refactor(backend): extract prompt helpers
